renamed scene package to level package which is a much better name (what was I thinking)
