+++ docs/GameCodeDetails/GameCodeDetailsSubSections/player.md
@@ -22,7 +22,7 @@ permalink: /GameCodeDetails/Player
 
 ## What does the Player class do?
 
-The `Player` class (found in the `Scene` package) is responsible for everything to do with the player (for this game, the player is the cat) 
+The `Player` class (found in the `Level` package) is responsible for everything to do with the player (for this game, the player is the cat) 
 during the platformer game except for the player character, which includes the player's graphics, movement, and interactions with the map (such as collision detection).
 
 In the gif below, everything the cat is doing and how it is interacting with the map is handled by the `Player` class.
