+++ src/Enemies/DinosaurEnemy.java
@@ -5,8 +5,8 @@ import Engine.ImageLoader;
 import GameObject.Frame;
 import GameObject.ImageEffect;
 import GameObject.SpriteSheet;
-import Scene.Enemy;
-import Scene.Player;
+import Level.Enemy;
+import Level.Player;
 import Utils.Point;
 import Utils.AirGroundState;
 import Utils.Direction;
