+++ src/GameObject/GameObject.java
@@ -2,8 +2,8 @@ package GameObject;
 
 import Builders.FrameBuilder;
 import Engine.GraphicsHandler;
-import Scene.Map;
-import Scene.MapTileCollisionHandler;
+import Level.Map;
+import Level.MapTileCollisionHandler;
 import Utils.Direction;
 import Utils.MathUtils;
 import java.awt.*;
