+++ src/Level/LevelState.java
@@ -1,4 +1,4 @@
-package Scene;
+package Level;
 
 public enum LevelState {
     RUNNING, LEVEL_COMPLETED, PLAYER_DEAD, LEVEL_WIN_MESSAGE, LEVEL_LOSE_MESSAGE
