+++ src/Level/MapEntity.java
@@ -1,4 +1,4 @@
-package Scene;
+package Level;
 
 import GameObject.*;
 import GameObject.Frame;
