+++ src/Level/MapEntityStatus.java
@@ -1,4 +1,4 @@
-package Scene;
+package Level;
 
 public enum MapEntityStatus {
     ACTIVE, INACTIVE, REMOVED
