+++ src/Level/NPC.java
@@ -1,4 +1,4 @@
-package Scene;
+package Level;
 
 import Engine.GraphicsHandler;
 import Engine.Key;
