+++ src/Level/Player.java
@@ -1,4 +1,4 @@
-package Scene;
+package Level;
 
 import Engine.Key;
 import Engine.KeyLocker;
