+++ src/Level/PlayerState.java
@@ -1,4 +1,4 @@
-package Scene;
+package Level;
 
 public enum PlayerState {
     STANDING, WALKING, JUMPING, CROUCHING
