+++ src/Level/TileType.java
@@ -1,4 +1,4 @@
-package Scene;
+package Level;
 
 public enum TileType {
     PASSABLE, NOT_PASSABLE, JUMP_THROUGH_PLATFORM
