+++ src/MapEditor/ChangeMapSizeWindow.java
@@ -1,7 +1,7 @@
 package MapEditor;
 
-import Scene.Map;
-import Scene.MapTile;
+import Level.Map;
+import Level.MapTile;
 import Utils.Colors;
 
 import javax.swing.*;
