+++ src/MapEditor/EditorMaps.java
@@ -2,10 +2,9 @@ package MapEditor;
 
 import Maps.TestMap;
 import Maps.TitleScreenMap;
-import Scene.Map;
+import Level.Map;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 
 public class EditorMaps {
     public static ArrayList<String> getMapNames() {
