+++ src/NPCs/Walrus.java
@@ -5,9 +5,9 @@ import Engine.GraphicsHandler;
 import Engine.ImageLoader;
 import GameObject.*;
 import GameObject.Frame;
-import Scene.Map;
-import Scene.NPC;
-import Scene.Player;
+import Level.Map;
+import Level.NPC;
+import Level.Player;
 import SpriteFont.SpriteFont;
 import Utils.Point;
 import java.awt.*;
